remove old events fetcher, and use new one
