+++ common/services/prismic/event-series.js
@@ -1,6 +1,5 @@
 // @flow
 import type {EventSeries} from '../../model/event-series';
-import type {UiEvent} from '../../model/events';
 import type {PrismicDocument, PrismicQueryOpts} from './types';
 import Prismic from 'prismic-javascript';
 import {getEvents} from './events';
@@ -38,7 +37,6 @@ export async function getEventSeries(req: ?Request, {
   const events = await getEvents(req, {
     page: 1,
     predicates: [Prismic.Predicates.at('my.events.series.series', id)],
-    order: 'asc',
     ...opts
   });
