+++ content/webapp/pages/events.js
@@ -1,15 +1,15 @@
 // @flow
 import type {Context} from 'next';
 import { Component } from 'react';
-import { getEvents } from '@weco/common/services/prismic/events';
+import { getEvents, orderEventsByNextAvailableDate } from '@weco/common/services/prismic/events';
 import { eventLd } from '@weco/common/utils/json-ld';
 import PageLayout from '@weco/common/views/components/PageLayout/PageLayout';
 import LayoutPaginatedResults from '@weco/common/views/components/LayoutPaginatedResults/LayoutPaginatedResults';
 import type { UiEvent } from '@weco/common/model/events';
 import type { PaginatedResults } from '@weco/common/services/prismic/types';
 import type { Period } from '@weco/common/model/periods';
-import {convertJsonToDates} from './event';
-import {convertImageUri} from '@weco/common/utils/convert-image-uri';
+import { convertImageUri } from '@weco/common/utils/convert-image-uri';
+import { convertJsonToDates } from './event';
 
 type Props = {|
   displayTitle: string,
@@ -21,12 +21,11 @@ const pageDescription = 'Choose from an inspiring range of free talks, tours, di
 export class ArticleSeriesPage extends Component<Props> {
   static getInitialProps = async (ctx: Context) => {
     const { page = 1 } = ctx.query;
-    const {period = 'current-and-coming-up'} = ctx.query;
+    const { period = 'current-and-coming-up' } = ctx.query;
     const events = await getEvents(ctx.req, {
       page,
       pageSize: 100,
-      period,
-      order: period === 'past' ? 'desc' : 'asc'
+      period
     });
     if (events && events.results.length > 0) {
       const title = (period === 'past' ? 'Past e' : 'E') + 'vents';
@@ -46,7 +45,7 @@ export class ArticleSeriesPage extends Component<Props> {
     const convertedEvents = events.results.map(convertJsonToDates);
     const convertedPaginatedResults = ({
       ...events,
-      results: convertedEvents
+      results: period !== 'past' ? orderEventsByNextAvailableDate(convertedEvents) : convertedEvents
     }: PaginatedResults<UiEvent>);
     const firstEvent = events.results[0];
